Rename starfish.typing to starfish.types
PyCharm confuses `import typing` in starfish.XXX as the relative module, even though that's not permitted in python 3.6.  This removes the false conflict.
